requirements.txt
# YouTube Downloader — Python dependencies
# Install with:  pip install -r requirements.txt

# Core download engine (actively maintained youtube-dl fork)
yt-dlp>=2024.3.10

# GUI framework
PyQt5>=5.15.9

# HTTP requests (thumbnail loading)
requests>=2.31.0

# Image processing (optional — for future thumbnail manipulation)
Pillow>=10.0.0
